Resolve "Refactor smart contract status return codes into a library"

Closes #22

See merge request sofie/open-source/offer-marketplace!14

+++ solidity/contracts/interfaces/ArrayExtraData.sol
@@ -2,7 +2,7 @@ pragma solidity ^0.5.8;
 
 
 interface ArrayExtraData {
-    function submitOfferArrayExtra(uint offerID, uint[] calldata extra) external returns (int status, uint offID);
+    function submitOfferArrayExtra(uint offerID, uint[] calldata extra) external returns (uint8 status, uint offID);
 
-    function submitRequestArrayExtra(uint requestID, uint[] calldata extra) external returns (int status, uint reqID);
+    function submitRequestArrayExtra(uint requestID, uint[] calldata extra) external returns (uint8 status, uint reqID);
 }
