+++ solidity/contracts/interfaces/StringExtraData.sol
@@ -2,7 +2,7 @@ pragma solidity ^0.5.8;
 
 
 interface JsonStringExtraData {
-    function submitOfferJsonStringExtra(uint offerID, string calldata extra) external returns (int status, uint offID);
+    function submitOfferJsonStringExtra(uint offerID, string calldata extra) external returns (uint8 status, uint offID);
 
-    function submitRequestJsonStringExtra(uint requestID, string calldata extra) external returns (int status, uint reqID);
+    function submitRequestJsonStringExtra(uint requestID, string calldata extra) external returns (uint8 status, uint reqID);
 }
